make ```ResourceContainer``` a decorator instead of class due to multiple metaclasses inheritance problem
